move hours functions to separate file
